+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 0.2.0 (unreleased)
 
+## Features
+
+* General `encode!` method that can encode a benchee structure to JSON - just delegates to Poison atm but decouples other plugins from it.
+
 ## Bugfixes
 
 * Relaxed `poison` dependency to be just `>= 1.4.0`